requirements.txt
# requirements.txt
# Проект: Система управления заказами типографии
# Автоматически сгенерировано

asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.7.1
cffi==2.0.0
channels-redis==4.1.0
channels==4.0.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.0.0
Django==4.2.7
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
packaging==25.0
psycopg2-binary==2.9.9
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
python-dotenv==1.0.0
redis==5.0.1
service-identity==24.2.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
zope.interface==8.1.1
django-mptt==0.14.0
# ===== ДОБАВЛЕНО ДЛЯ ЭКСПОРТА В EXCEL =====
# openpyxl – библиотека для чтения/записи файлов Excel (.xlsx).
# Позволяет создавать полноценные Excel-документы с форматированием.
openpyxl==3.1.2